fix(email): Total Mailboxes was concatenating counts, not summing them (#651)
`count(*)` is bigint, and node-postgres returns bigint as a STRING to avoid
precision loss. `mailboxCount` was typed `number` and arrived as `"3"`, so the
Email header tile's `reduce((sum, d) => sum + (d.mailboxCount ?? 0), 0)` did
string concatenation: with production's sixteen mail domains it rendered a
~17-digit number where the answer is 34. The sortable Mailboxes column had the
matching defect — lexicographic ordering puts "10" before "2".

This is the defect actually behind the "wrong number" report. The earlier commit
in this release fixed the tile's SEMANTICS (it counted the platform's own
postmaster@ intake mailboxes); this fixes its ARITHMETIC. Both were real, and
only the second one was visible as nonsense.

Found by calling the deployed endpoint on DEV and looking at the JSON rather
than trusting the declared type: the field is annotated `sql<number>` and every
test fixture in the suite used a number, so nothing in the repo disagreed with
itself.

Fixed at the source — `::int` on all three count subqueries, so every consumer
gets a real number and sorting behaves — plus `Number()` at the tile, because a
`number`-typed field arriving as text has now happened once and the tile should
not be the place that discovers it again.

The new tests use the shape the API really sends (`"3"`, not `3`) and pin the
exact pre-fix rendering: reverting the coercion fails with
`expected '0324' to be '9'`.

+++ backend/src/modules/email-domains/service.ts
@@ -651,7 +651,7 @@ export async function getEmailDomain(
   // own plumbing on the tenant's domain, not mailboxes the tenant has or can
   // see. See the subqueries in listEmailDomains / listAllEmailDomains.
   const [mailboxCount] = await db
-    .select({ count: sql<number>`count(*)` })
+    .select({ count: sql<number>`count(*)::int` })
     .from(mailboxes)
     .where(and(
       eq(mailboxes.emailDomainId, emailDomain.id),
@@ -791,7 +791,15 @@ export async function listEmailDomains(
       // on production, 50 against the 34 mailboxes an operator can actually
       // open, the extra 16 being the postmaster@ report-intake boxes the
       // platform provisions for itself.
-      mailboxCount: sql<number>`(SELECT count(*) FROM mailboxes WHERE mailboxes.email_domain_id = ${emailDomains.id} AND mailboxes.platform_managed = false)`,
+      // `::int` is load-bearing, not tidiness. `count(*)` is bigint, and node-postgres
+      // hands bigint back as a STRING to avoid precision loss — so this field was
+      // typed `number` and arrived as `"3"`. The Email header tile sums it with
+      // `reduce((sum, d) => sum + (d.mailboxCount ?? 0), 0)`, and in JS
+      // `0 + "3"` is `"03"`: with production's 16 domains the "Total Mailboxes"
+      // number rendered as a ~17-digit concatenation of the per-domain counts.
+      // The sortable Mailboxes column had the matching defect — string ordering
+      // puts "10" before "2". A mailbox count can never overflow int4.
+      mailboxCount: sql<number>`(SELECT count(*) FROM mailboxes WHERE mailboxes.email_domain_id = ${emailDomains.id} AND mailboxes.platform_managed = false)::int`,
     })
     .from(emailDomains)
     .innerJoin(domains, eq(emailDomains.domainId, domains.id))
@@ -827,7 +835,15 @@ export async function listAllEmailDomains(db: Database) {
       // on production, 50 against the 34 mailboxes an operator can actually
       // open, the extra 16 being the postmaster@ report-intake boxes the
       // platform provisions for itself.
-      mailboxCount: sql<number>`(SELECT count(*) FROM mailboxes WHERE mailboxes.email_domain_id = ${emailDomains.id} AND mailboxes.platform_managed = false)`,
+      // `::int` is load-bearing, not tidiness. `count(*)` is bigint, and node-postgres
+      // hands bigint back as a STRING to avoid precision loss — so this field was
+      // typed `number` and arrived as `"3"`. The Email header tile sums it with
+      // `reduce((sum, d) => sum + (d.mailboxCount ?? 0), 0)`, and in JS
+      // `0 + "3"` is `"03"`: with production's 16 domains the "Total Mailboxes"
+      // number rendered as a ~17-digit concatenation of the per-domain counts.
+      // The sortable Mailboxes column had the matching defect — string ordering
+      // puts "10" before "2". A mailbox count can never overflow int4.
+      mailboxCount: sql<number>`(SELECT count(*) FROM mailboxes WHERE mailboxes.email_domain_id = ${emailDomains.id} AND mailboxes.platform_managed = false)::int`,
     })
     .from(emailDomains)
     .innerJoin(domains, eq(emailDomains.domainId, domains.id));
